Exercicis presentacio 2

+++ AA2_1_Exercicis/Exercici_6.sql
@@ -10,7 +10,7 @@ CREATE TABLE zoo
 	ciutat VARCHAR(20),
 	pais VARCHAR(20),
 	pressupost DOUBLE,
-	grandaria DOUBLE,
+	grandaria INT,
 	PRIMARY KEY (id_zoo),
 	CONSTRAINT chk_zoo CHECK (grandaria BETWEEN 1000 AND 1000000)
 );
